@@ -1,17 +1,24 @@
 [package]
 name = "rustlings"
-version = "4.5.0"
-authors = ["anastasie <ana@ana.st>", "Carol (Nichols || Goulding) <carol.nichols@gmail.com>"]
-edition = "2018"
+description = "Small exercises to get you used to reading and writing Rust code!"
+version = "5.5.1"
+authors = [
+  "Liv <mokou@fastmail.com>",
+  "Carol (Nichols || Goulding) <carol.nichols@gmail.com>",
+]
+edition = "2021"
 
 [dependencies]
-argh = "0.1.4"
-indicatif = "0.10.3"
-console = "0.7.7"
-notify = "4.0.15"
-toml = "0.4.10"
-regex = "1.1.6"
-serde = {version = "1.0.10", features = ["derive"]}
+argh = "0.1"
+indicatif = "0.16"
+console = "0.15"
+notify = "4.0"
+toml = "0.5"
+regex = "1.5"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.81"
+home = "0.5.3"
+glob = "0.3.0"
 
 [[bin]]
 name = "rustlings"
