@@ -1,5 +1,6 @@
 use crate::exercise::{Exercise, ExerciseList};
-use crate::run::run;
+use crate::project::RustAnalyzerProject;
+use crate::run::{reset, run};
 use crate::verify::verify;
 use argh::FromArgs;
 use console::Emoji;
@@ -10,7 +11,8 @@ use std::fs;
 use std::io::{self, prelude::*};
 use std::path::Path;
 use std::process::{Command, Stdio};
-use std::sync::mpsc::channel;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::mpsc::{channel, RecvTimeoutError};
 use std::sync::{Arc, Mutex};
 use std::thread;
 use std::time::Duration;
@@ -19,11 +21,12 @@ use std::time::Duration;
 mod ui;
 
 mod exercise;
+mod project;
 mod run;
 mod verify;
 
 // In sync with crate version
-const VERSION: &str = "4.5.0";
+const VERSION: &str = "5.5.1";
 
 #[derive(FromArgs, PartialEq, Debug)]
 /// Rustlings is a collection of small exercises to get you used to writing and reading Rust code
@@ -44,8 +47,10 @@ enum Subcommands {
     Verify(VerifyArgs),
     Watch(WatchArgs),
     Run(RunArgs),
+    Reset(ResetArgs),
     Hint(HintArgs),
     List(ListArgs),
+    Lsp(LspArgs),
 }
 
 #[derive(FromArgs, PartialEq, Debug)]
@@ -56,12 +61,25 @@ struct VerifyArgs {}
 #[derive(FromArgs, PartialEq, Debug)]
 #[argh(subcommand, name = "watch")]
 /// Reruns `verify` when files were edited
-struct WatchArgs {}
+struct WatchArgs {
+    /// show hints on success
+    #[argh(switch)]
+    success_hints: bool,
+}
 
 #[derive(FromArgs, PartialEq, Debug)]
 #[argh(subcommand, name = "run")]
 /// Runs/Tests a single exercise
 struct RunArgs {
+    #[argh(positional)]
+    /// the name of the exercise
+    name: String,
+}
+
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(subcommand, name = "reset")]
+/// Resets a single exercise using "git stash -- <filename>"
+struct ResetArgs {
     #[argh(positional)]
     /// the name of the exercise
     name: String,
@@ -75,6 +93,11 @@ struct HintArgs {
     /// the name of the exercise
     name: String,
 }
+
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(subcommand, name = "lsp")]
+/// Enable rust-analyzer for exercises
+struct LspArgs {}
 
 #[derive(FromArgs, PartialEq, Debug)]
 #[argh(subcommand, name = "list")]
@@ -102,20 +125,12 @@ fn main() {
     let args: Args = argh::from_env();
 
     if args.version {
-        println!("v{}", VERSION);
+        println!("v{VERSION}");
         std::process::exit(0);
     }
 
     if args.nested.is_none() {
-        println!();
-        println!(r#"       welcome to...                      "#);
-        println!(r#"                 _   _ _                  "#);
-        println!(r#"  _ __ _   _ ___| |_| (_)_ __   __ _ ___  "#);
-        println!(r#" | '__| | | / __| __| | | '_ \ / _` / __| "#);
-        println!(r#" | |  | |_| \__ \ |_| | | | | | (_| \__ \ "#);
-        println!(r#" |_|   \__,_|___/\__|_|_|_| |_|\__, |___/ "#);
-        println!(r#"                               |___/      "#);
-        println!();
+        println!("\n{WELCOME}\n");
     }
 
     if !Path::new("info.toml").exists() {
@@ -139,8 +154,7 @@ fn main() {
     let verbose = args.nocapture;
 
     let command = args.nested.unwrap_or_else(|| {
-        let text = fs::read_to_string("default_out.txt").unwrap();
-        println!("{}", text);
+        println!("{DEFAULT_OUT}\n");
         std::process::exit(0);
     });
     match command {
@@ -169,11 +183,11 @@ fn main() {
                 };
                 if solve_cond && (filter_cond || subargs.filter.is_none()) {
                     let line = if subargs.paths {
-                        format!("{}\n", fname)
+                        format!("{fname}\n")
                     } else if subargs.names {
                         format!("{}\n", e.name)
                     } else {
-                        format!("{:<17}\t{:<46}\t{:<7}\n", e.name, fname, status)
+                        format!("{:<17}\t{fname:<46}\t{status:<7}\n", e.name)
                     };
                     // Somehow using println! leads to the binary panicking
                     // when its output is piped.
@@ -192,7 +206,7 @@ fn main() {
             });
             let percentage_progress = exercises_done as f32 / exercises.len() as f32 * 100.0;
             println!(
-                "Progress: You completed {} / {} exercises ({:.2} %).",
+                "Progress: You completed {} / {} exercises ({:.1} %).",
                 exercises_done,
                 exercises.len(),
                 percentage_progress
@@ -206,6 +220,12 @@ fn main() {
             run(exercise, verbose).unwrap_or_else(|_| std::process::exit(1));
         }
 
+        Subcommands::Reset(subargs) => {
+            let exercise = find_exercise(&subargs.name, &exercises);
+
+            reset(exercise).unwrap_or_else(|_| std::process::exit(1));
+        }
+
         Subcommands::Hint(subargs) => {
             let exercise = find_exercise(&subargs.name, &exercises);
 
@@ -213,11 +233,31 @@ fn main() {
         }
 
         Subcommands::Verify(_subargs) => {
-            verify(&exercises, verbose).unwrap_or_else(|_| std::process::exit(1));
+            verify(&exercises, (0, exercises.len()), verbose, false)
+                .unwrap_or_else(|_| std::process::exit(1));
         }
 
-        Subcommands::Watch(_subargs) => {
-            if let Err(e) = watch(&exercises, verbose) {
+        Subcommands::Lsp(_subargs) => {
+            let mut project = RustAnalyzerProject::new();
+            project
+                .get_sysroot_src()
+                .expect("Couldn't find toolchain path, do you have `rustc` installed?");
+            project
+                .exercises_to_json()
+                .expect("Couldn't parse rustlings exercises files");
+
+            if project.crates.is_empty() {
+                println!("Failed find any exercises, make sure you're in the `rustlings` folder");
+            } else if project.write_to_disk().is_err() {
+                println!("Failed to write rust-project.json to disk for rust-analyzer");
+            } else {
+                println!("Successfully generated rust-project.json");
+                println!("rust-analyzer will now parse exercises, restart your language server or editor")
+            }
+        }
+
+        Subcommands::Watch(_subargs) => match watch(&exercises, verbose, _subargs.success_hints) {
+            Err(e) => {
                 println!(
                     "Error: Could not watch your progress. Error message was {:?}.",
                     e
@@ -225,62 +265,63 @@ fn main() {
                 println!("Most likely you've run out of disk space or your 'inotify limit' has been reached.");
                 std::process::exit(1);
             }
-            println!(
-                "{emoji} All exercises completed! {emoji}",
-                emoji = Emoji("🎉", "★")
-            );
-            println!();
-            println!("+----------------------------------------------------+");
-            println!("|          You made it to the Fe-nish line!          |");
-            println!("+--------------------------  ------------------------+");
-            println!("                          \\/                         ");
-            println!("     ▒▒          ▒▒▒▒▒▒▒▒      ▒▒▒▒▒▒▒▒          ▒▒   ");
-            println!("   ▒▒▒▒  ▒▒    ▒▒        ▒▒  ▒▒        ▒▒    ▒▒  ▒▒▒▒ ");
-            println!("   ▒▒▒▒  ▒▒  ▒▒            ▒▒            ▒▒  ▒▒  ▒▒▒▒ ");
-            println!(" ░░▒▒▒▒░░▒▒  ▒▒            ▒▒            ▒▒  ▒▒░░▒▒▒▒ ");
-            println!("   ▓▓▓▓▓▓▓▓  ▓▓      ▓▓██  ▓▓  ▓▓██      ▓▓  ▓▓▓▓▓▓▓▓ ");
-            println!("     ▒▒▒▒    ▒▒      ████  ▒▒  ████      ▒▒░░  ▒▒▒▒   ");
-            println!("       ▒▒  ▒▒▒▒▒▒        ▒▒▒▒▒▒        ▒▒▒▒▒▒  ▒▒     ");
-            println!("         ▒▒▒▒▒▒▒▒▒▒▓▓▓▓▓▓▒▒▒▒▒▒▒▒▓▓▒▒▓▓▒▒▒▒▒▒▒▒       ");
-            println!("           ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒         ");
-            println!("             ▒▒▒▒▒▒▒▒▒▒██▒▒▒▒▒▒██▒▒▒▒▒▒▒▒▒▒           ");
-            println!("           ▒▒  ▒▒▒▒▒▒▒▒▒▒██████▒▒▒▒▒▒▒▒▒▒  ▒▒         ");
-            println!("         ▒▒    ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒    ▒▒       ");
-            println!("       ▒▒    ▒▒    ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒    ▒▒    ▒▒     ");
-            println!("       ▒▒  ▒▒    ▒▒                  ▒▒    ▒▒  ▒▒     ");
-            println!("           ▒▒  ▒▒                      ▒▒  ▒▒         ");
-            println!();
-            println!("We hope you enjoyed learning about the various aspects of Rust!");
-            println!(
-                "If you noticed any issues, please don't hesitate to report them to our repo."
-            );
-            println!("You can also contribute your own exercises to help the greater community!");
-            println!();
-            println!("Before reporting an issue or contributing, please read our guidelines:");
-            println!("https://github.com/rust-lang/rustlings/blob/main/CONTRIBUTING.md");
-        }
+            Ok(WatchStatus::Finished) => {
+                println!(
+                    "{emoji} All exercises completed! {emoji}",
+                    emoji = Emoji("🎉", "★")
+                );
+                println!("\n{FENISH_LINE}\n");
+            }
+            Ok(WatchStatus::Unfinished) => {
+                println!("We hope you're enjoying learning about Rust!");
+                println!("If you want to continue working on the exercises at a later point, you can simply run `rustlings watch` again");
+            }
+        },
     }
 }
 
-fn spawn_watch_shell(failed_exercise_hint: &Arc<Mutex<Option<String>>>) {
+fn spawn_watch_shell(
+    failed_exercise_hint: &Arc<Mutex<Option<String>>>,
+    should_quit: Arc<AtomicBool>,
+) {
     let failed_exercise_hint = Arc::clone(failed_exercise_hint);
-    println!("Type 'hint' or open the corresponding README.md file to get help or type 'clear' to clear the screen.");
+    println!("Welcome to watch mode! You can type 'help' to get an overview of the commands you can use here.");
     thread::spawn(move || loop {
         let mut input = String::new();
         match io::stdin().read_line(&mut input) {
             Ok(_) => {
                 let input = input.trim();
-                if input.eq("hint") {
+                if input == "hint" {
                     if let Some(hint) = &*failed_exercise_hint.lock().unwrap() {
-                        println!("{}", hint);
+                        println!("{hint}");
                     }
-                } else if input.eq("clear") {
+                } else if input == "clear" {
                     println!("\x1B[2J\x1B[1;1H");
+                } else if input.eq("quit") {
+                    should_quit.store(true, Ordering::SeqCst);
+                    println!("Bye!");
+                } else if input.eq("help") {
+                    println!("Commands available to you in watch mode:");
+                    println!("  hint   - prints the current exercise's hint");
+                    println!("  clear  - clears the screen");
+                    println!("  quit   - quits watch mode");
+                    println!("  !<cmd> - executes a command, like `!rustc --explain E0381`");
+                    println!("  help   - displays this help message");
+                    println!();
+                    println!("Watch mode automatically re-evaluates the current exercise");
+                    println!("when you edit a file's contents.")
+                } else if let Some(cmd) = input.strip_prefix('!') {
+                    let parts: Vec<&str> = cmd.split_whitespace().collect();
+                    if parts.is_empty() {
+                        println!("no command provided");
+                    } else if let Err(e) = Command::new(parts[0]).args(&parts[1..]).status() {
+                        println!("failed to execute command `{}`: {}", cmd, e);
+                    }
                 } else {
-                    println!("unknown command: {}", input);
+                    println!("unknown command: {input}");
                 }
             }
-            Err(error) => println!("error reading command: {}", error),
+            Err(error) => println!("error reading command: {error}"),
         }
     });
 }
@@ -300,13 +341,22 @@ fn find_exercise<'a>(name: &str, exercises: &'a [Exercise]) -> &'a Exercise {
             .iter()
             .find(|e| e.name == name)
             .unwrap_or_else(|| {
-                println!("No exercise found for '{}'!", name);
+                println!("No exercise found for '{name}'!");
                 std::process::exit(1)
             })
     }
 }
 
-fn watch(exercises: &[Exercise], verbose: bool) -> notify::Result<()> {
+enum WatchStatus {
+    Finished,
+    Unfinished,
+}
+
+fn watch(
+    exercises: &[Exercise],
+    verbose: bool,
+    success_hints: bool,
+) -> notify::Result<WatchStatus> {
     /* Clears the terminal with an ANSI escape code.
     Works in UNIX and newer Windows terminals. */
     fn clear_screen() {
@@ -314,36 +364,48 @@ fn watch(exercises: &[Exercise], verbose: bool) -> notify::Result<()> {
     }
 
     let (tx, rx) = channel();
+    let should_quit = Arc::new(AtomicBool::new(false));
 
-    let mut watcher: RecommendedWatcher = Watcher::new(tx, Duration::from_secs(2))?;
+    let mut watcher: RecommendedWatcher = Watcher::new(tx, Duration::from_secs(1))?;
     watcher.watch(Path::new("./exercises"), RecursiveMode::Recursive)?;
 
     clear_screen();
 
     let to_owned_hint = |t: &Exercise| t.hint.to_owned();
-    let failed_exercise_hint = match verify(exercises.iter(), verbose) {
-        Ok(_) => return Ok(()),
+    let failed_exercise_hint = match verify(
+        exercises.iter(),
+        (0, exercises.len()),
+        verbose,
+        success_hints,
+    ) {
+        Ok(_) => return Ok(WatchStatus::Finished),
         Err(exercise) => Arc::new(Mutex::new(Some(to_owned_hint(exercise)))),
     };
-    spawn_watch_shell(&failed_exercise_hint);
+    spawn_watch_shell(&failed_exercise_hint, Arc::clone(&should_quit));
     loop {
-        match rx.recv() {
+        match rx.recv_timeout(Duration::from_secs(1)) {
             Ok(event) => match event {
                 DebouncedEvent::Create(b) | DebouncedEvent::Chmod(b) | DebouncedEvent::Write(b) => {
                     if b.extension() == Some(OsStr::new("rs")) && b.exists() {
                         let filepath = b.as_path().canonicalize().unwrap();
                         let pending_exercises = exercises
                             .iter()
-                            .skip_while(|e| !filepath.ends_with(&e.path))
-                            // .filter(|e| filepath.ends_with(&e.path))
+                            .find(|e| filepath.ends_with(&e.path))
+                            .into_iter()
                             .chain(
                                 exercises
                                     .iter()
                                     .filter(|e| !e.looks_done() && !filepath.ends_with(&e.path)),
                             );
+                        let num_done = exercises.iter().filter(|e| e.looks_done()).count();
                         clear_screen();
-                        match verify(pending_exercises, verbose) {
-                            Ok(_) => return Ok(()),
+                        match verify(
+                            pending_exercises,
+                            (num_done, exercises.len()),
+                            verbose,
+                            success_hints,
+                        ) {
+                            Ok(_) => return Ok(WatchStatus::Finished),
                             Err(exercise) => {
                                 let mut failed_exercise_hint = failed_exercise_hint.lock().unwrap();
                                 *failed_exercise_hint = Some(to_owned_hint(exercise));
@@ -353,7 +415,14 @@ fn watch(exercises: &[Exercise], verbose: bool) -> notify::Result<()> {
                 }
                 _ => {}
             },
-            Err(e) => println!("watch error: {:?}", e),
+            Err(RecvTimeoutError::Timeout) => {
+                // the timeout expired, just check the `should_quit` variable below then loop again
+            }
+            Err(e) => println!("watch error: {e:?}"),
+        }
+        // Check if we need to exit
+        if should_quit.load(Ordering::SeqCst) {
+            return Ok(WatchStatus::Unfinished);
         }
     }
 }
@@ -367,3 +436,66 @@ fn rustc_exists() -> bool {
         .map(|status| status.success())
         .unwrap_or(false)
 }
+
+const DEFAULT_OUT: &str = r#"Thanks for installing Rustlings!
+
+Is this your first time? Don't worry, Rustlings was made for beginners! We are
+going to teach you a lot of things about Rust, but before we can get
+started, here's a couple of notes about how Rustlings operates:
+
+1. The central concept behind Rustlings is that you solve exercises. These
+   exercises usually have some sort of syntax error in them, which will cause
+   them to fail compilation or testing. Sometimes there's a logic error instead
+   of a syntax error. No matter what error, it's your job to find it and fix it!
+   You'll know when you fixed it because then, the exercise will compile and
+   Rustlings will be able to move on to the next exercise.
+2. If you run Rustlings in watch mode (which we recommend), it'll automatically
+   start with the first exercise. Don't get confused by an error message popping
+   up as soon as you run Rustlings! This is part of the exercise that you're
+   supposed to solve, so open the exercise file in an editor and start your
+   detective work!
+3. If you're stuck on an exercise, there is a helpful hint you can view by typing
+   'hint' (in watch mode), or running `rustlings hint exercise_name`.
+4. If an exercise doesn't make sense to you, feel free to open an issue on GitHub!
+   (https://github.com/rust-lang/rustlings/issues/new). We look at every issue,
+   and sometimes, other learners do too so you can help each other out!
+5. If you want to use `rust-analyzer` with exercises, which provides features like
+   autocompletion, run the command `rustlings lsp`.
+
+Got all that? Great! To get started, run `rustlings watch` in order to get the first
+exercise. Make sure to have your editor open!"#;
+
+const FENISH_LINE: &str = r#"+----------------------------------------------------+
+|          You made it to the Fe-nish line!          |
++--------------------------  ------------------------+
+                          \\/
+     ▒▒          ▒▒▒▒▒▒▒▒      ▒▒▒▒▒▒▒▒          ▒▒
+   ▒▒▒▒  ▒▒    ▒▒        ▒▒  ▒▒        ▒▒    ▒▒  ▒▒▒▒
+   ▒▒▒▒  ▒▒  ▒▒            ▒▒            ▒▒  ▒▒  ▒▒▒▒
+ ░░▒▒▒▒░░▒▒  ▒▒            ▒▒            ▒▒  ▒▒░░▒▒▒▒
+   ▓▓▓▓▓▓▓▓  ▓▓      ▓▓██  ▓▓  ▓▓██      ▓▓  ▓▓▓▓▓▓▓▓
+     ▒▒▒▒    ▒▒      ████  ▒▒  ████      ▒▒░░  ▒▒▒▒
+       ▒▒  ▒▒▒▒▒▒        ▒▒▒▒▒▒        ▒▒▒▒▒▒  ▒▒
+         ▒▒▒▒▒▒▒▒▒▒▓▓▓▓▓▓▒▒▒▒▒▒▒▒▓▓▒▒▓▓▒▒▒▒▒▒▒▒
+           ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
+             ▒▒▒▒▒▒▒▒▒▒██▒▒▒▒▒▒██▒▒▒▒▒▒▒▒▒▒
+           ▒▒  ▒▒▒▒▒▒▒▒▒▒██████▒▒▒▒▒▒▒▒▒▒  ▒▒
+         ▒▒    ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒    ▒▒
+       ▒▒    ▒▒    ▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒    ▒▒    ▒▒
+       ▒▒  ▒▒    ▒▒                  ▒▒    ▒▒  ▒▒
+           ▒▒  ▒▒                      ▒▒  ▒▒
+
+We hope you enjoyed learning about the various aspects of Rust!
+If you noticed any issues, please don't hesitate to report them to our repo.
+You can also contribute your own exercises to help the greater community!
+
+Before reporting an issue or contributing, please read our guidelines:
+https://github.com/rust-lang/rustlings/blob/main/CONTRIBUTING.md"#;
+
+const WELCOME: &str = r#"       welcome to...
+                 _   _ _
+  _ __ _   _ ___| |_| (_)_ __   __ _ ___
+ | '__| | | / __| __| | | '_ \ / _` / __|
+ | |  | |_| \__ \ |_| | | | | | (_| \__ \
+ |_|   \__,_|___/\__|_|_|_| |_|\__, |___/
+                               |___/"#;
