@@ -1,5 +1,9 @@
 // enums3.rs
+//
 // Address all the TODOs to make the tests pass!
+//
+// Execute `rustlings hint enums3` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
@@ -16,6 +20,7 @@ struct State {
     color: (u8, u8, u8),
     position: Point,
     quit: bool,
+    message: String
 }
 
 impl State {
@@ -27,16 +32,17 @@ impl State {
         self.quit = true;
     }
 
-    fn echo(&self, s: String) {
-        println!("{}", s);
-    }
+    fn echo(&mut self, s: String) { self.message = s }
 
     fn move_position(&mut self, p: Point) {
         self.position = p;
     }
 
     fn process(&mut self, message: Message) {
-        // TODO: create a match expression to process the different message variants
+        // TODO: create a match expression to process the different message
+        // variants
+        // Remember: When passing a tuple as a function argument, you'll need
+        // extra parentheses: fn function((t, u, p, l, e))
     }
 }
 
@@ -50,8 +56,9 @@ mod tests {
             quit: false,
             position: Point { x: 0, y: 0 },
             color: (0, 0, 0),
+            message: "hello world".to_string(),
         };
-        state.process(Message::ChangeColor((255, 0, 255)));
+        state.process(Message::ChangeColor(255, 0, 255));
         state.process(Message::Echo(String::from("hello world")));
         state.process(Message::Move(Point { x: 10, y: 15 }));
         state.process(Message::Quit);
@@ -60,5 +67,6 @@ mod tests {
         assert_eq!(state.position.x, 10);
         assert_eq!(state.position.y, 15);
         assert_eq!(state.quit, true);
+        assert_eq!(state.message, "hello world");
     }
 }
