@@ -1,19 +1,22 @@
 // errors1.rs
-// This function refuses to generate text to be printed on a nametag if
-// you pass it an empty string. It'd be nicer if it explained what the problem
-// was, instead of just sometimes returning `None`. The 2nd test currently
-// does not compile or pass, but it illustrates the behavior we would like
-// this function to have.
-// Execute `rustlings hint errors1` for hints!
+//
+// This function refuses to generate text to be printed on a nametag if you pass
+// it an empty string. It'd be nicer if it explained what the problem was,
+// instead of just sometimes returning `None`. Thankfully, Rust has a similar
+// construct to `Result` that can be used to express error conditions. Let's use
+// it!
+//
+// Execute `rustlings hint errors1` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
 pub fn generate_nametag_text(name: String) -> Option<String> {
-    if name.len() > 0 {
-        Some(format!("Hi! My name is {}", name))
-    } else {
+    if name.is_empty() {
         // Empty names aren't allowed.
         None
+    } else {
+        Some(format!("Hi! My name is {}", name))
     }
 }
 
@@ -21,14 +24,11 @@ pub fn generate_nametag_text(name: String) -> Option<String> {
 mod tests {
     use super::*;
 
-    // This test passes initially if you comment out the 2nd test.
-    // You'll need to update what this test expects when you change
-    // the function under test!
     #[test]
     fn generates_nametag_text_for_a_nonempty_name() {
         assert_eq!(
             generate_nametag_text("Beyoncé".into()),
-            Some("Hi! My name is Beyoncé".into())
+            Ok("Hi! My name is Beyoncé".into())
         );
     }
 
@@ -36,6 +36,7 @@ mod tests {
     fn explains_why_generating_nametag_text_fails() {
         assert_eq!(
             generate_nametag_text("".into()),
+            // Don't change this line
             Err("`name` was empty; it must be nonempty.".into())
         );
     }
