@@ -1,5 +1,7 @@
 // functions2.rs
-// Make me compile! Execute `rustlings hint functions2` for hints :)
+//
+// Execute `rustlings hint functions2` or use the `hint` watch subcommand for a
+// hint.
 
 fn main() {
     call_me(3);
