@@ -1,5 +1,7 @@
 // functions3.rs
-// Make me compile! Execute `rustlings hint functions3` for hints :)
+//
+// Execute `rustlings hint functions3` or use the `hint` watch subcommand for a
+// hint.
 
 fn main() {
     call_me(5);
