@@ -1,7 +1,10 @@
+// generics2.rs
+//
 // This powerful wrapper provides the ability to store a positive integer value.
 // Rewrite it using generics so that it supports wrapping ANY type.
-
-// Execute `rustlings hint generics2` for hints!
+//
+// Execute `rustlings hint generics2` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
