@@ -1,10 +1,12 @@
 // if2.rs
-
+//
 // Step 1: Make me compile!
 // Step 2: Get the bar_for_fuzz and default_to_baz tests passing!
-// Execute the command `rustlings hint if2` if you want a hint :)
+//
+// Execute `rustlings hint if2` or use the `hint` watch subcommand for a hint.
 
-pub fn fizz_if_foo(fizzish: &str) -> &str {
+
+pub fn foo_if_fizz(fizzish: &str) -> &str {
     if fizzish == "fizz" {
         "foo"
     } else if fizzish == "fuzz" {
@@ -21,16 +23,16 @@ mod tests {
 
     #[test]
     fn foo_for_fizz() {
-        assert_eq!(fizz_if_foo("fizz"), "foo")
+        assert_eq!(foo_if_fizz("fizz"), "foo")
     }
 
     #[test]
     fn bar_for_fuzz() {
-        assert_eq!(fizz_if_foo("fuzz"), "bar")
+        assert_eq!(foo_if_fizz("fuzz"), "bar")
     }
 
     #[test]
     fn default_to_baz() {
-        assert_eq!(fizz_if_foo("literally anything"), "baz")
+        assert_eq!(foo_if_fizz("literally anything"), "baz")
     }
 }
