@@ -1,5 +1,7 @@
 // macros1.rs
-// Make me compile! Execute `rustlings hint macros1` for hints :)
+//
+// Execute `rustlings hint macros1` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
