@@ -1,5 +1,7 @@
 // macros2.rs
-// Make me compile! Execute `rustlings hint macros2` for hints :)
+//
+// Execute `rustlings hint macros2` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
