@@ -1,6 +1,9 @@
 // macros3.rs
+//
 // Make me compile, without taking the macro out of the module!
-// Execute `rustlings hint macros3` for hints :)
+//
+// Execute `rustlings hint macros3` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
