@@ -1,6 +1,11 @@
 // move_semantics2.rs
-// Make me compile without changing line 13!
-// Execute `rustlings hint move_semantics2` for hints :)
+//
+// Expected output:
+// vec0 has length 3, with contents `[22, 44, 66]`
+// vec1 has length 4, with contents `[22, 44, 66, 88]`
+//
+// Execute `rustlings hint move_semantics2` or use the `hint` watch subcommand
+// for a hint.
 
 
 fn main() {
@@ -8,12 +13,11 @@ fn main() {
 
     let mut vec1 = fill_vec(&vec0);
 
-    // Do not change the following line!
-    println!("{} has length {} content `{:?}`", "vec0", vec0.len(), vec0);
+    println!("{} has length {}, with contents: `{:?}`", "vec0", vec0.len(), vec0);
 
     vec1.push(88);
 
-    println!("{} has length {} content `{:?}`", "vec1", vec1.len(), vec1);
+    println!("{} has length {}, with contents `{:?}`", "vec1", vec1.len(), vec1);
 }
 
 fn fill_vec(vec: &Vec<i32>) -> Vec<i32> {
