@@ -1,12 +1,18 @@
 // primitive_types2.rs
-// Fill in the rest of the line that has code missing!
-// No hints, there's no tricks, just get used to typing these :)
+//
+// Fill in the rest of the line that has code missing! No hints, there's no
+// tricks, just get used to typing these :)
+//
+// Execute `rustlings hint primitive_types2` or use the `hint` watch subcommand
+// for a hint.
 
 // I AM NOT DONE
 
 fn main() {
     // Characters (`char`)
 
+    // Note the _single_ quotes, these are different from the double quotes
+    // you've been seeing around.
     let my_first_initial = 'C';
     if my_first_initial.is_alphabetic() {
         println!("Alphabetical!");
