@@ -1,6 +1,9 @@
 // primitive_types3.rs
+//
 // Create an array with at least 100 elements in it where the ??? is.
-// Execute `rustlings hint primitive_types3` for hints!
+//
+// Execute `rustlings hint primitive_types3` or use the `hint` watch subcommand
+// for a hint.
 
 // I AM NOT DONE
 
