@@ -1,6 +1,9 @@
 // primitive_types5.rs
+//
 // Destructure the `cat` tuple so that the println will work.
-// Execute `rustlings hint primitive_types5` for hints!
+//
+// Execute `rustlings hint primitive_types5` or use the `hint` watch subcommand
+// for a hint.
 
 // I AM NOT DONE
 
