@@ -1,6 +1,9 @@
 // strings2.rs
+//
 // Make me compile without changing the function signature!
-// Execute `rustlings hint strings2` for hints :)
+//
+// Execute `rustlings hint strings2` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
