@@ -1,5 +1,9 @@
 // structs1.rs
+//
 // Address all the TODOs to make the tests pass!
+//
+// Execute `rustlings hint structs1` or use the `hint` watch subcommand for a
+// hint.
 
 // I AM NOT DONE
 
@@ -10,7 +14,7 @@ struct ColorClassicStruct {
 struct ColorTupleStruct(/* TODO: Something goes here */);
 
 #[derive(Debug)]
-struct UnitStruct;
+struct UnitLikeStruct;
 
 #[cfg(test)]
 mod tests {
@@ -21,8 +25,9 @@ mod tests {
         // TODO: Instantiate a classic c struct!
         // let green =
 
-        assert_eq!(green.name, "green");
-        assert_eq!(green.hex, "#00FF00");
+        assert_eq!(green.red, 0);
+        assert_eq!(green.green, 255);
+        assert_eq!(green.blue, 0);
     }
 
     #[test]
@@ -30,16 +35,17 @@ mod tests {
         // TODO: Instantiate a tuple struct!
         // let green =
 
-        assert_eq!(green.0, "green");
-        assert_eq!(green.1, "#00FF00");
+        assert_eq!(green.0, 0);
+        assert_eq!(green.1, 255);
+        assert_eq!(green.2, 0);
     }
 
     #[test]
     fn unit_structs() {
-        // TODO: Instantiate a unit struct!
-        // let unit_struct =
-        let message = format!("{:?}s are fun!", unit_struct);
+        // TODO: Instantiate a unit-like struct!
+        // let unit_like_struct =
+        let message = format!("{:?}s are fun!", unit_like_struct);
 
-        assert_eq!(message, "UnitStructs are fun!");
+        assert_eq!(message, "UnitLikeStructs are fun!");
     }
 }
