@@ -1,14 +1,12 @@
 // traits2.rs
 //
-// Your task is to implement the trait
-// `AppendBar' for a vector of strings.
-//
-// To implement this trait, consider for
-// a moment what it means to 'append "Bar"'
+// Your task is to implement the trait `AppendBar` for a vector of strings. To
+// implement this trait, consider for a moment what it means to 'append "Bar"'
 // to a vector of strings.
 //
-// No boiler plate code this time,
-// you can do this!
+// No boiler plate code this time, you can do this!
+//
+// Execute `rustlings hint traits2` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
@@ -16,7 +14,7 @@ trait AppendBar {
     fn append_bar(self) -> Self;
 }
 
-//TODO: Add your code here
+// TODO: Implement trait `AppendBar` for a vector of strings.
 
 #[cfg(test)]
 mod tests {
