@@ -1,9 +1,10 @@
 // variables3.rs
-// Make me compile! Execute the command `rustlings hint variables3` if you want a hint :)
+//
+// Execute `rustlings hint variables3` or use the `hint` watch subcommand for a
+// hint.
+
 
 fn main() {
-    let mut x = 3;
-    println!("Number {}", x);
-    x = 5; // don't change this line
+    let mut x: i32 = 10;
     println!("Number {}", x);
 }
