@@ -1,7 +1,12 @@
 // variables4.rs
-// Make me compile! Execute the command `rustlings hint variables4` if you want a hint :)
+//
+// Execute `rustlings hint variables4` or use the `hint` watch subcommand for a
+// hint.
+
 
 fn main() {
-    let x: i32 = 10;
+    let mut x = 3;
+    println!("Number {}", x);
+    x = 5; // don't change this line
     println!("Number {}", x);
 }
