@@ -1,5 +1,8 @@
 // variables5.rs
-// Make me compile! Execute the command `rustlings hint variables5` if you want a hint :)
+//
+// Execute `rustlings hint variables5` or use the `hint` watch subcommand for a
+// hint.
+
 
 fn main() {
     let number = "T-H-R-E-E"; // don't change this line
