@@ -1,5 +1,7 @@
 // variables6.rs
-// Make me compile! Execute the command `rustlings hint variables6` if you want a hint :)
+//
+// Execute `rustlings hint variables6` or use the `hint` watch subcommand for a
+// hint.
 
 const NUMBER: u8 = 3;
 fn main() {
